+++ projectbarber/src/main/java/com/projectbarber/projectbarber/repository/BarbeiroRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.projectbarber.projectbarber.model.Barbeiro;
 
-public interface BarbeiroRepository extends JpaRepository<Barbeiro, Long> {
+public interface BarbeiroRepository extends JpaRepository<Barbeiro, Integer> {
     Optional<Barbeiro> findByEmail(String email);    
     Optional<Barbeiro> findByNome(String nome);
